+++ scripts/updateBlocks.js
@@ -28,6 +28,8 @@ const blocktomllink = 'https://raw.githubusercontent.com/scpwiki/wikijump/develo
       'module',
       'include-messy',
       'html',
+      'iframe',
+      'embed',
       'css',
       'math',
       'code',